Change error generated files to mmu format
