--- 04_MMU/button_operations.h
@@ -1,18 +1,0 @@
-#pragma once
-#include <stdint.h>
-
-namespace MMU2 {
-
-/// Will be mapped onto dialog button responses in the FW
-/// Those responses have their unique+translated texts as well
-enum class ButtonOperations : uint8_t {
-    NoOperation = 0,
-    Retry = 1,
-    Continue = 2,
-    ResetMMU = 3,
-    Unload = 4,
-    StopPrint = 5,
-    DisableMMU = 6,
-};
-
-} // namespace MMU2
