+++ include/error_type.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <stdint.h>
+
+enum class ErrType : uint8_t {
+    ERROR = 0,
+    WARNING,
+    USER_ACTION
+};
